WhisperRecorder/scripts/requirements.txt
# Speaker Diarization Dependencies for WhisperRecorder
# Install with: pip install -r requirements.txt

# Core pyannote.audio and dependencies
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0

# Additional dependencies that might be needed
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
librosa>=0.9.0
soundfile>=0.10.0

# Optional: For faster processing with GPU
# torch-audio-cuda  # Uncomment if you have CUDA GPU

# HuggingFace for model downloads
transformers>=4.21.0
huggingface-hub>=0.10.0

# Progress and logging
tqdm>=4.64.0

# Disable SpeechBrain temporarily due to sentencepiece build issues on macOS
# Will use pyannote.audio without SpeechBrain integration
# speechbrain>=1.0.0  # Commented out due to macOS build issues 